Split activation.rs up to add frame.rs
